create fcrash.h
